smart_shopper/requirements.txt
streamlit>=1.32
google-generativeai>=0.7
google-search-results>=2.4
pillow>=10.0
huggingface_hub>=0.23
requests>=2.31
